requirements.txt
torch
torchvision
Pillow
werkzeug
